cora/cora/requirements.txt
mss
google-generativeai
pystray
Pillow
PyQt6
ollama
keyboard
SpeechRecognition
pyaudio
